requirements.txt
# Core dependencies
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# ML Libraries
lightgbm>=3.0.0
xgboost>=1.5.0
catboost>=1.0.0
optuna>=3.0.0

# Kaggle API
kaggle>=1.5.12

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
joblib>=1.1.0
click>=8.0.0

# Optional: Data validation
# pandera>=0.13.0